fixed by linking every linked page evry other link
So in each page I put event listeners for every other one, so in 'HOME' I have events listeners for contact and menu as well as their imports, in 'CONTACT' as well for home and menu and so on... It's not the best thou.
