+++ src/contact.js
@@ -1,4 +1,6 @@
-
+import navbar from'./navbar'
+import menu from './menu'
+import home from './home'
 const contact = () =>{
     const content = document.getElementById('content');
     
@@ -26,7 +28,17 @@ const contact = () =>{
                         </div>                            
                     </div> `;
     content.innerHTML += template;
-    
+
+    const about = document.getElementById('about')
+    about.addEventListener('click', () => {
+        navbar();
+        home();
+    });
+    const menus = document.getElementById('menu')
+    menus.addEventListener('click', () => {
+        navbar();
+        menu();
+    });
     
 }
 export default contact;
