+++ src/index.js
@@ -1,6 +1,6 @@
 import navbar from './navbar';
 import home from './home';
-import menu from './menu'
+
 
 navbar();
 home();
