+++ src/menu.js
@@ -1,4 +1,6 @@
-
+import home from './home'
+import navbar from './navbar'
+import contact from './contact'
 const menu = () =>{
     const content = document.getElementById('content');
     
@@ -85,7 +87,16 @@ const menu = () =>{
                         </div>                            
                     </div> `;
     content.innerHTML += template;
-    
+    const about = document.getElementById('about')
+    about.addEventListener('click', () => {
+        navbar();
+        home();
+    });
+    const cont = document.getElementById('contact')
+    cont.addEventListener('click', () => {
+        navbar();
+        contact();
+    });
     
 }
 export default menu
