+++ src/navbar.js
@@ -1,6 +1,3 @@
-import menu from './menu'
-import home from  './home'
-import contact from './contact'
 const navbar = () =>{
     const content = document.getElementById('content');
     
@@ -29,21 +26,8 @@ const navbar = () =>{
                         `;
     content.innerHTML += template;
     
-    const about = document.getElementById('about')
-    about.addEventListener('click', () => {
-        navbar();
-        home();
-    });
-    const menus = document.getElementById('menu')
-    menus.addEventListener('click', () => {
-        navbar();
-        menu();
-    });
-    const cont = document.getElementById('contact')
-    cont.addEventListener('click', () => {
-        navbar();
-        contact();
-    });
+    
+    
 };
 
 export default navbar;
